requirements.txt
# Core
pyyaml>=6.0.1
httpx>=0.28.0
websockets>=12.0
pydantic>=2.7
aiosqlite>=0.20.0

# Optional (live trading / signing)
solana==0.36.9          # Note: requires httpx>=0.28.0
solders>=0.21.0
base58>=2.1.1

# Dev / tests
pytest>=8.2
pytest-asyncio>=0.23.8

# Additional
aiohttp>=3.9.5
